pyproject.toml
[project]
name = "pdf2md-ocr"
version = "0.0.3"
description = "Simple CLI tool to convert PDFs to Markdown using Marker AI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Carlos Castillo", email = "castillo.st+pdf2md-ocr@gmail.com" },
]
keywords = ["pdf", "markdown", "ocr", "conversion", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = ["marker-pdf==1.10.1", "click", "rich"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
pdf2md-ocr = "pdf2md_ocr.cli:main"

[project.urls]
Homepage = "https://github.com/carloscasalar/pdf2md-ocr"
Repository = "https://github.com/carloscasalar/pdf2md-ocr"
Issues = "https://github.com/carloscasalar/pdf2md-ocr/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf2md_ocr"]
